+++ packages/google_maps_flutter/google_maps_flutter/lib/src/google_map.dart
@@ -128,6 +128,7 @@ class GoogleMap extends StatefulWidget {
     this.onCameraIdle,
     this.onTap,
     this.onLongPress,
+    this.onPoiTap,
     this.cloudMapId,
   });
 
@@ -251,6 +252,9 @@ class GoogleMap extends StatefulWidget {
 
   /// Called every time a [GoogleMap] is long pressed.
   final ArgumentCallback<LatLng>? onLongPress;
+
+  /// Called every time a [PointOfInterest] is clicked.
+  final ArgumentCallback<PointOfInterest>? onPoiTap;
 
   /// True if a "My Location" layer should be shown on the map.
   ///
@@ -575,6 +579,14 @@ class _GoogleMapState extends State<GoogleMap> {
     final ArgumentCallback<LatLng>? onLongPress = widget.onLongPress;
     if (onLongPress != null) {
       onLongPress(position);
+    }
+  }
+
+  void onPoiTap(PointOfInterest poi) {
+    assert(poi != null);
+    final ArgumentCallback<PointOfInterest>? onPoiTap = widget.onPoiTap;
+    if (onPoiTap != null) {
+      onPoiTap(poi);
     }
   }